refactor: utilise shared duration formatter across plex/subsonic
- Migrated the contents of the plex duration function towards a new generalised utility function.
- Refactored subsonic logic to utilise shared duration function.
- Refactored plex towards this new function aswell.
